requirements.txt
black==22.8.0
cryptography==36.0.1
dash==2.15.0
dash-bootstrap-components==1.1.0
dash-extensions==1.0.1
dash-iconify==0.1.2
dash_mantine_components==0.14.3
detect-secrets==1.4.0
Flask==2.1.1
gunicorn==20.1.0
ipykernel==6.19.2
itsdangerous==2.1.2
nbclient==0.7.2
nbconvert==7.2.6
nbformat==5.7.0
numpy==1.22.4
pandas==1.5.0
plotly==5.7.0
python-dotenv==0.13.0
werkzeug==2.3.7